+++ tests/test_client.py
@@ -17,7 +17,6 @@ from requests.exceptions import ReadTimeout
 from six.moves import mock
 
 from reportportal_client.helpers import timestamp
-from reportportal_client.static.defines import NOT_FOUND
 
 
 def connection_error(*args, **kwargs):
@@ -31,24 +30,31 @@ def response_error(*args, **kwargs):
     return result
 
 
+def invalid_response(*args, **kwargs):
+    result = Response()
+    result._content = \
+        '<html><head><title>405 Not Allowed</title></head></html>'
+    result.status_code = 405
+    return result
+
+
 @pytest.mark.parametrize(
-    'requests_method, client_method, client_params, expected_result',
+    'requests_method, client_method, client_params',
     [
-        ('put', 'finish_launch', [timestamp()], NOT_FOUND),
-        ('put', 'finish_test_item', ['test_item_id', timestamp()], NOT_FOUND),
-        ('get', 'get_item_id_by_uuid', ['test_item_uuid'], NOT_FOUND),
-        ('get', 'get_launch_info', [], {}),
-        ('get', 'get_launch_ui_id', [], None),
-        ('get', 'get_launch_ui_url', [], None),
-        ('get', 'get_project_settings', [], {}),
-        ('post', 'start_launch', ['Test Launch', timestamp()], NOT_FOUND),
-        ('post', 'start_test_item', ['Test Item', timestamp(), 'STEP'],
-         NOT_FOUND),
-        ('put', 'update_test_item', ['test_item_id'], NOT_FOUND)
+        ('put', 'finish_launch', [timestamp()]),
+        ('put', 'finish_test_item', ['test_item_id', timestamp()]),
+        ('get', 'get_item_id_by_uuid', ['test_item_uuid']),
+        ('get', 'get_launch_info', []),
+        ('get', 'get_launch_ui_id', []),
+        ('get', 'get_launch_ui_url', []),
+        ('get', 'get_project_settings', []),
+        ('post', 'start_launch', ['Test Launch', timestamp()]),
+        ('post', 'start_test_item', ['Test Item', timestamp(), 'STEP']),
+        ('put', 'update_test_item', ['test_item_id'])
     ]
 )
 def test_connection_errors(rp_client, requests_method, client_method,
-                           client_params, expected_result):
+                           client_params):
     rp_client.launch_id = 'test_launch_id'
     getattr(rp_client.session, requests_method).side_effect = connection_error
     result = getattr(rp_client, client_method)(*client_params)
@@ -56,7 +62,30 @@ def test_connection_errors(rp_client, requests_method, client_method,
 
     getattr(rp_client.session, requests_method).side_effect = response_error
     result = getattr(rp_client, client_method)(*client_params)
-    assert result == expected_result
+    assert result is None
+
+
+@pytest.mark.parametrize(
+    'requests_method, client_method, client_params',
+    [
+        ('put', 'finish_launch', [timestamp()]),
+        ('put', 'finish_test_item', ['test_item_id', timestamp()]),
+        ('get', 'get_item_id_by_uuid', ['test_item_uuid']),
+        ('get', 'get_launch_info', []),
+        ('get', 'get_launch_ui_id', []),
+        ('get', 'get_launch_ui_url', []),
+        ('get', 'get_project_settings', []),
+        ('post', 'start_launch', ['Test Launch', timestamp()]),
+        ('post', 'start_test_item', ['Test Item', timestamp(), 'STEP']),
+        ('put', 'update_test_item', ['test_item_id'])
+    ]
+)
+def test_invalid_responses(rp_client, requests_method, client_method,
+                           client_params):
+    rp_client.launch_id = 'test_launch_id'
+    getattr(rp_client.session, requests_method).side_effect = invalid_response
+    result = getattr(rp_client, client_method)(*client_params)
+    assert result is None
 
 
 LAUNCH_ID = 333